refactor compositie_solution
