+++ solvis_graphql_api/composite_solution/__init__.py
@@ -5,5 +5,6 @@ from .composite_rupture_detail import (  # SortRupturesArgs,
     SimpleSortRupturesArgs,
 )
 from .composite_rupture_sections import CompositeRuptureSections
-from .composite_solution import CompositeSolution, FilterRupturesArgs, FilterRupturesArgsInput
+from .composite_solution import CompositeSolution
+from .filtered_ruptures_args import FilterRupturesArgs, FilterRupturesArgsInput
 from .schema import paginated_filtered_ruptures
